requirements.txt
django==4.0.6
djangorestframework==3.13.1
drf_yasg==1.21.3
flake8==4.0.1
psycopg2==2.9.3
pylint==2.14.5
python-dotenv==0.20.0
